pyproject.toml
[project]
name = "ashybulakstroy-mcp-1c-bridge"
version = "0.8.0"
description = "AshybulakStroy MCP 1C Bridge: MCP server for safe AI access to 1C Accounting Kazakhstan 3.0 via OData"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.6.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
ashybulak-1c-bridge = "ashybulakstroy_mcp_1c_bridge.mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
